requirements_stylegan.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.3.0=mkl
absl-py=0.15.0=pyhd3eb1b0_0
asgiref=3.5.0=pypi_0
astor=0.8.1=py37h06a4308_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
cached-property=1.5.2=py_0
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py37h89c1867_1
charset-normalizer=2.0.7=pypi_0
click=8.0.3=pypi_0
cudatoolkit=11.1.1=h6406543_8
cython=0.29.26=pypi_0
django=3.2.12=pypi_0
fbpca=1.0=pypi_0
ffmpeg=1.4=pypi_0
ffmpeg-python=0.2.0=pypi_0
flatbuffers=2.0=pypi_0
freetype=2.11.0=h70c0345_0
future=0.18.2=pypi_0
gast=0.2.2=py37_0
giflib=5.2.1=h7b6447c_0
glumpy=1.2.0=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.43.0=pypi_0
h5py=3.6.0=py37ha0f2276_0
hdf5=1.10.6=hb1b8bf9_0
idna=3.3=pypi_0
imageio=2.13.2=pypi_0
imageio-ffmpeg=0.4.3=pypi_0
importlib-metadata=4.8.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9d=h7f8727e_0
keras=2.7.0=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libclang=12.0.0=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
lmdb=1.3.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py37h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
ninja=1.10.2.2=pypi_0
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py37_0
opencv-python=4.5.5.62=pypi_0
openh264=2.1.1=h4ff587b_0
opensimplex=0.4=pypi_0
openssl=1.1.1m=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
pillow=8.4.0=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
protobuf=3.19.3=pypi_0
psutil=5.8.0=py37h5e8e339_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyopengl=3.1.5=pypi_0
pyopengltk=0.0.3=pypi_0
pyspng=0.1.0=pypi_0
python=3.7.11=h12debd9_0
python_abi=3.7=2_cp37m
pytorch=1.10.2=py3.7_cuda11.1_cudnn8.0.5_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.8=pypi_0
scipy=1.7.3=py37hc147768_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.2=hc218d9a_0
sqlparse=0.4.2=pypi_0
tensorboard=1.15.0=py37_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=1.15.0=mkl_py37h28c19af_0
tensorflow-base=1.15.0=mkl_py37he1670d9_0
tensorflow-estimator=1.15.1=pyh2649769_0
tensorflow-gpu=1.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.23.1=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch-utils=0.1.2=pypi_0
torchaudio=0.10.2=py37_cu111
torchvision=0.11.3=py37_cu111
tqdm=4.62.3=pypi_0
triangle=20200424=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.13.3=py37h7f8727e_2
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
